Move `ModalComponentTree` to `components.tree` package
